% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DecisionTrees.R
\name{dtPrediction}
\alias{dtPrediction}
\title{Decision Trees Classifier}
\usage{
dtPrediction(train.data.set, test.data.set, labels = c("Region"),
  categorical.var = c("SampleId"), cp = 0.001)
}
\description{
Decision Trees Classifier
}
