% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetData.R
\name{explodeGDS}
\alias{explodeGDS}
\title{Explode GDS}
\usage{
explodeGDS(gds, output.gds.gn, times)
}
\arguments{
\item{gds}{input gds}

\item{times}{times to expolode}

\item{output.gds.fn}{output gds}
}
\description{
Explode the gds X times
Use function snpgdsCombineGeno from SNPRelate
}
