% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetData.R
\name{getPCAdf}
\alias{getPCAdf}
\title{Get reduced DataFrame}
\usage{
getPCAdf(pca, n.pc, pop.df)
}
\arguments{
\item{pca}{a result PCA DataFrame}

\item{n.pc}{a number of PCs}

\item{pop.df}{a samples panel}
}
\value{
a DataFrame after PCA
}
\description{
Combine DF from PCA with samples panel DF. Not exported
}
