% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetData.R
\name{getQCdf}
\alias{getQCdf}
\title{Get DataFrame}
\usage{
getQCdf(genofile, pop.df, min.maf = 0.005, max.maf = 0.05, ld = 0.2,
  missing.rate = 0)
}
\arguments{
\item{genofile}{the .gds file}

\item{pop.df}{the samples panel}

\item{min.maf}{the inf limit of minor allele frequencies}

\item{max.maf}{the sup limit of minor allele frequencies}

\item{ld}{the linkage disequlibrium threshold}

\item{missing.rate}{a missiang rate threshold}
}
\value{
a DataFrame after PCA
}
\description{
Perform quality control analysis of the given genotype and return DataFrame to be fed to ML models
}
