% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetData.R
\name{getQCgds}
\alias{getQCgds}
\title{Get gds after QC}
\usage{
getQCgds(genofile, pop.df, min.maf = 0.005, max.maf = 0.05, ld = 0.2,
  missing.rate = 0)
}
\value{
The gds object after QC
}
\description{
Get gds after QC
}
