% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetData.R
\name{getQCPCAdf}
\alias{getQCPCAdf}
\title{Get reduced DataFrame}
\usage{
getQCPCAdf(genofile, pop.df, min.maf = 0.005, max.maf = 0.05, ld = 0.2,
  missing.rate = 0, npc = 4, with.scree.plot = F, with.eig.tbl = F)
}
\arguments{
\item{genofile}{the .gds file}

\item{pop.df}{the samples panel}

\item{min.maf}{the inf limit of minor allele frequencies}

\item{max.maf}{the sup limit of minor allele frequencies}

\item{ld}{the linkage disequlibrium threshold}

\item{missing.rate}{a missiang rate threshold}

\item{npc}{the number of principal components}

\item{with.scree.plot}{if TRUE return the scree plot of the first 15th PCs, default = FALSE}

\item{with.eig.tbl}{if TRUE return the table with eigenvalues and variances of obtained  PCs, default = FALSE}
}
\value{
a DataFrame after PCA
}
\description{
Perform quality control analysis and dimensionality reduction of the given genotype and return DataFrame to be fed to ML models
}
