% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{panelLoad}
\alias{panelLoad}
\title{Load a list of individuals for an example data}
\usage{
panelLoad()
}
\value{
The a list of individuals
}
\description{
Load a sample panel from 1000 Genomes Project
}
