% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomForest.R
\name{rfPrediction}
\alias{rfPrediction}
\title{Random Forest Classifier}
\usage{
rfPrediction(train.data.set, test.data.set, labels = c("Region"),
  categorical.var = c("SampleId"), ntree = 500)
}
\description{
Random Forest Classifier
}
