% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryClustering.R
\name{summaryClustering}
\alias{summaryClustering}
\title{Summary for clustering evaluation metrics}
\usage{
summaryClustering(obj.hclust, obj.kmeans, obj.emc)
}
\description{
Summary for clustering evaluation metrics
}
