% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVM.R
\name{svmEstimation}
\alias{svmEstimation}
\title{SVM  Estimation}
\usage{
svmEstimation(train.data.set, labels = c("Region"),
  categorical.var = c("SampleId"), which.kernel = "RBF", costs = c(10,
  100, 1000, 10000), sigmas = 10^(-2:1))
}
\description{
SVM  Estimation
}
