% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVM.R
\name{svmPrediction}
\alias{svmPrediction}
\title{SVM prediction}
\usage{
svmPrediction(train.data.set, test.data.set, labels = c("Region"),
  categorical.var = c("SampleId"), which.kernel = "RBF", cost = 100,
  sigma = 1)
}
\description{
SVM prediction
}
