% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TuningForClustering.R
\name{tuningLD}
\alias{tuningLD}
\title{Tuning LD for clustering}
\usage{
tuningLD(genofile, pop.df, min.maf, max.maf, ld.set = ld.set,
  missing.rate = 0, n.pc = 4, n.clusters = 5)
}
\arguments{
\item{genofile}{gds file}

\item{pop.df}{panel file that contains labels, e.g. Region}

\item{min.maf}{infimum maf for filtering SNPs}

\item{max.maf}{supremum maf for filtering SNPs}

\item{ld.set}{linkage-disequilibrium range for tuning}

\item{n.pc}{number of principal components (PCA)}

\item{n.clusters}{number of clusters}
}
\description{
Tuning LD for clustering
}
