% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TuningForClassification.R
\name{tuningLDclass}
\alias{tuningLDclass}
\title{Tuning LD for classification}
\usage{
tuningLDclass(genofile, pop.df, min.maf, max.maf, ld.set, missing.rate = 0,
  n.pc = 5, labels = c("Region"), categorical.var = c("SampleId"),
  split.coef = 0.67, cost = 10, sigma = 1, cp = 0.001, ntree = 500)
}
\arguments{
\item{cost}{for svm}

\item{sigma}{for all kernels if svm apart of linear one}

\item{cp}{control parameter for decision trees}

\item{ntree}{number of trees for random forest}

\item{split}{split df into train and test}
}
\description{
Tuning LD for classification
}
