% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TuningForClustering.R
\name{tuningLdMaf}
\alias{tuningLdMaf}
\title{Tuning MAF vs  LD}
\usage{
tuningLdMaf(genofile, pop.df, inf.maf.set, sup.maf.set, ld.set = ld.set,
  missing.rate = 0, n.pc = 4, n.clusters = 5, method = "hierarchical")
}
\description{
Tuning MAF vs  LD
}
