% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TuningForClustering.R
\name{tuningLDnPC}
\alias{tuningLDnPC}
\title{Tuning LD vs  nPC}
\usage{
tuningLDnPC(genofile, pop.df, min.maf, max.maf, ld.set, n.pc.set,
  missing.rate = 0, n.clusters = 5, method = "hierarchical")
}
\description{
Tuning LD vs  nPC
}
