% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TuningForClassification.R
\name{tuningLDnPCclass}
\alias{tuningLDnPCclass}
\title{Tuning LD vs nPC for classification}
\usage{
tuningLDnPCclass(genofile, pop.df, min.maf, max.maf, ld.set, n.pc.set,
  missing.rate = 0, labels = c("Region"), categorical.var = c("SampleId"),
  split.coef = 0.67, cost = 10, sigma = 1, cp = 0.001, ntree = 500,
  method = "svml")
}
\arguments{
\item{genofile}{gds file}

\item{pop.df}{panel file that contains labels, e.g. Region}

\item{min.maf}{infimum maf for filtering SNPs}

\item{max.maf}{supremum maf for filtering SNPs}

\item{ld.set}{linkage-disequilibrium range for tuning}

\item{cost}{for svm}

\item{sigma}{for all kernels if svm apart of linear one}

\item{cp}{control parameter for decision trees}

\item{ntree}{number of trees for random forest}

\item{method}{classification method}

\item{n.pc}{set of numbers of principal components to tune}

\item{split}{split df into train and test}
}
\description{
Tuning LD vs nPC for classification
}
\details{
With method = 'svml' function tunes parameters for SVM with linear kernel, 'svmq' - for SVM with quadratic kernel, 'svmr' - for SVM with radial kernel, 'dt' - for Decision Trees, 'rfs' - for Random Forest
}
