% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TuningForClustering.R
\name{tuningMAF}
\alias{tuningMAF}
\title{Tuning MAF for clusterig}
\usage{
tuningMAF(genofile, pop.df, inf.maf.set, sup.maf.set, ld = 0.2,
  missing.rate = 0, n.pc = 4, n.clusters = 5)
}
\arguments{
\item{genofile}{gds file}

\item{pop.df}{panel file that contains labels, e.g. Region}

\item{inf.maf.set}{range of infimum maf for tuning}

\item{sup.maf.set}{range of supremum maf for tuning}

\item{ld}{linkage-disequilibrium threshold}

\item{n.pc}{number of principal components (PCA)}

\item{n.clusters}{number of clusters}
}
\description{
Tuning MAF for clusterig
}
