% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TuningForClassification.R
\name{tuningMAFclass}
\alias{tuningMAFclass}
\title{Tuning MAF for classification}
\usage{
tuningMAFclass(genofile, pop.df, inf.maf.set, sup.maf.set, ld = 0.2,
  missing.rate = 0, n.pc = 5, labels = c("Region"),
  categorical.var = c("SampleId"), split.coef = 0.67, cost = 10,
  sigma = 1, cp = 0.001, ntree = 500)
}
\arguments{
\item{genofile}{gds file}

\item{pop.df}{panel file that contains labels, e.g. Region}

\item{inf.maf.set}{range of infimum maf for tuning}

\item{sup.maf.set}{range of supremum maf for tuning}

\item{ld}{linkage-disequilibrium threshold}

\item{n.pc}{number of principal components (PCA)}

\item{cost}{for svm}

\item{sigma}{for all kernels if svm apart of linear one}

\item{cp}{control parameter for decision trees}

\item{ntree}{number of trees for random forest}

\item{split}{split df into train and test}
}
\description{
Tuning MAF for classification
}
