% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TuningForClustering.R
\name{tuningMAFPC}
\alias{tuningMAFPC}
\title{Tuning MAF vs  nPC}
\usage{
tuningMAFPC(genofile, pop.df, inf.maf.set, sup.maf.set, n.pc.set, ld = 0.2,
  missing.rate = 0, n.clusters = 5, method = "hierarchical")
}
\description{
Tuning MAF vs  nPC
}
