% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TuningForClustering.R
\name{tuningNumPC}
\alias{tuningNumPC}
\title{Tuning nPC for clustering}
\usage{
tuningNumPC(genofile, pop.df, min.maf, max.maf, ld, n.pc.set,
  missing.rate = 0, n.clusters = 5)
}
\arguments{
\item{genofile}{gds file}

\item{pop.df}{panel file that contains labels, e.g. Region}

\item{min.maf}{infimum maf for filtering SNPs}

\item{max.maf}{supremum maf for filtering SNPs}

\item{ld}{linkage-disequilibrium threshold}

\item{n.clusters}{number of clusters}

\item{n.pc}{set of numbers of principal components to tune}
}
\description{
Tuning nPC for clustering
}
