% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TuningForClassification.R
\name{tuningNumPCclass}
\alias{tuningNumPCclass}
\title{Tuning nPC for classification}
\usage{
tuningNumPCclass(genofile, pop.df, min.maf, max.maf, ld, n.pc.set,
  missing.rate = 0, labels = c("Region"), categorical.var = c("SampleId"),
  split.coef = 0.67, cost = 10, sigma = 1, cp = 0.001, ntree = 500)
}
\arguments{
\item{cost}{for svm}

\item{sigma}{for all kernels if svm apart of linear one}

\item{cp}{control parameter for decision trees}

\item{ntree}{number of trees for random forest}

\item{split}{split df into train and test}
}
\description{
Tuning nPC for classification
}
