% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TuningForClustering.R
\name{tuningLdMaf}
\alias{tuningLdMaf}
\title{Tuning MAF vs  LD for clustering}
\usage{
tuningLdMaf(genofile, pop.df, inf.maf.set, sup.maf.set, ld.set = ld.set,
  missing.rate = 0, n.pc = 4, n.clusters = 5, method = "hierarchical")
}
\arguments{
\item{genofile}{gds file}

\item{pop.df}{panel file that contains labels, e.g. Region}

\item{inf.maf.set}{range of infimum maf for tuning}

\item{sup.maf.set}{range of supremum maf for tuning}

\item{ld.set}{linkage-disequilibrium range for tuning}

\item{n.pc}{number of principal components (PCA)}

\item{n.clusters}{number of clusters}

\item{method}{clustering method}
}
\description{
Tuning MAF vs  LD for clustering
}
\details{
Three option for "method" = {'hierarchical', 'kmeans', 'em'}
}
