% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TuningForClustering.R
\name{tuningLDnPC}
\alias{tuningLDnPC}
\title{Tuning LD vs  nPC for clustering}
\usage{
tuningLDnPC(genofile, pop.df, min.maf, max.maf, ld.set, n.pc.set,
  missing.rate = 0, n.clusters = 5, method = "hierarchical")
}
\arguments{
\item{genofile}{gds file}

\item{pop.df}{panel file that contains labels, e.g. Region}

\item{min.maf}{infimum maf for filtering SNPs}

\item{max.maf}{supremum maf for filtering SNPs}

\item{ld.set}{linkage-disequilibrium range for tuning}

\item{n.clusters}{number of clusters}

\item{method}{clustering method}

\item{n.pc}{set of numbers of principal components to tune}
}
\description{
Tuning LD vs  nPC for clustering
}
\details{
Three option for "method" = {'hierarchical', 'kmeans', 'em'}
}
