% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TuningForClustering.R
\name{tuningMAF}
\alias{tuningMAF}
\title{Tuning MAF}
\usage{
tuningMAF(genofile, pop.df, inf.maf.set, sup.maf.set, ld = 0.2,
  missing.rate = 0, n.pc = 4, n.clusters = 5)
}
\description{
Tuning MAF
}
