% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TuningForClustering.R
\name{tuningMAFPC}
\alias{tuningMAFPC}
\title{Tuning MAF vs  nPC for clustering}
\usage{
tuningMAFPC(genofile, pop.df, inf.maf.set, sup.maf.set, n.pc.set, ld = 0.2,
  missing.rate = 0, n.clusters = 5, method = "hierarchical")
}
\arguments{
\item{genofile}{gds file}

\item{pop.df}{panel file that contains labels, e.g. Region}

\item{inf.maf.set}{range of infimum maf for tuning}

\item{sup.maf.set}{range of supremum maf for tuning}

\item{ld}{linkage-disequilibrium threshold}

\item{n.clusters}{number of clusters}

\item{method}{clustering method}

\item{n.pc}{set of numbers of principal components to tune}
}
\description{
Tuning MAF vs  nPC for clustering
}
\details{
Three option for "method" = {'hierarchical', 'kmeans', 'em'}
}
